requirements.txt
python-dotenv
discord
discord.py
requests
ratelimit